+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.8.0 (2025-07-01)
+
+- Added `purge` and `purge_all` API for purging snaps.
+- Added `list_all` API for listing all snaps.
+- Added `snapshots` and `forget_snapshots` for listing and forgetting snapshotsd
+
+
 ## 1.7.0 (2025-06-27)
 
 - Added `get_recovery_systems` API for listing snapd recovery systems.